Added fleeing and Line-of-Sight detection
Entities will now abandon their wandering and pick a 'hiding spot' object in the scene to path to when a player is detected. In order for a player to be detected, they must be both within the sight radius of the entity, and have a clear line of sight (which is calculated using a raycast.)

Once they have picked a spot to flee to, they path there using Unity's AI navigation system, and a timer starts counting down. Once the timer counts down to 0, they go back to their regular wandering - the idea being that if you're chasing multiple entities, and one runs away behind a corner while you chase another one, it'll eventually go back to where you first found it.

It also works perfectly fine for multiple enemies, though there's currently no way to keep multiple entities from pathing to the same hiding spot.