Stealth boost can now be activated, but having trouble with new inventory errors.
Stealth boost is now functional in playermovement, but I am seeing a new error with the inventory where the noggin item keeps triggering a null error when attempting to remove it from the inventory. This is not happening with the other items, so not sure why. Maybe has something to do with the buff?

I also removed the weird duplicate items.