Added stamina boost, inventory error definitely seems related to boosts
Stamina boost works, but now the error is associated with the leg too, which means its definitely the boosts causing it. I'm still not entirely sure why it's happening cause it's only when I first try to use the item upon opening the inventory. If I close the inventory, reopen it, and try to use the item again, it works fine. It's only ever the first opening of the inventory that doesn't work.